removed the legacy apikeys screen keeping only the new one
