+++ apps/studio/pages/org/[slug]/project/[ref]/branch/[branch]/settings/api-keys/index.tsx
@@ -1,14 +1,38 @@
+import {
+  ApiKeysCreateCallout,
+  ApiKeysFeedbackBanner,
+} from 'components/interfaces/APIKeys/ApiKeysIllustrations'
+import { useApiKeysVisibility } from 'components/interfaces/APIKeys/hooks/useApiKeysVisibility'
+import { PublishableAPIKeys } from 'components/interfaces/APIKeys/PublishableAPIKeys'
+import { SecretAPIKeys } from 'components/interfaces/APIKeys/SecretAPIKeys'
 import ApiKeysLayout from 'components/layouts/APIKeys/APIKeysLayout'
 import DefaultLayout from 'components/layouts/DefaultLayout'
 import SettingsLayout from 'components/layouts/ProjectSettingsLayout/SettingsLayout'
-import { DisplayApiSettings } from 'components/ui/ProjectSettings'
+import { DisableInteraction } from 'components/ui/DisableInteraction'
 import type { NextPageWithLayout } from 'types'
+import { Separator } from 'ui'
 
-const ApiKeysLegacyPage: NextPageWithLayout = () => {
-  return <DisplayApiSettings showTitle={false} showNotice={false} />
+// const ApiKeysLegacyPage: NextPageWithLayout = () => {
+//   return <DisplayApiSettings showTitle={false} showNotice={false} />
+// }
+
+const ApiKeysNewPage: NextPageWithLayout = () => {
+  const { shouldDisableUI, canInitApiKeys } = useApiKeysVisibility()
+
+  return (
+    <>
+      {canInitApiKeys && <ApiKeysCreateCallout />}
+      <ApiKeysFeedbackBanner />
+      <DisableInteraction disabled={shouldDisableUI} className="flex flex-col gap-8">
+        <PublishableAPIKeys />
+        <Separator />
+        <SecretAPIKeys />
+      </DisableInteraction>
+    </>
+  )
 }
 
-ApiKeysLegacyPage.getLayout = (page) => (
+ApiKeysNewPage.getLayout = (page) => (
   <DefaultLayout>
     <SettingsLayout>
       <ApiKeysLayout>{page}</ApiKeysLayout>
@@ -16,4 +40,4 @@ ApiKeysLegacyPage.getLayout = (page) => (
   </DefaultLayout>
 )
 
-export default ApiKeysLegacyPage
+export default ApiKeysNewPage
